[Modify] OrderedImg Container / Presenter 분리 #261
